1. feat(accounts): support CPA JSON imports with persistent fingerprints
- accept structured account imports with token, plan type, and fingerprint metadata
- map CPA JSON fingerprint data into persisted account fp fields
- keep stable device/session identifiers for imported accounts
- preserve imported paid account type when refresh data reports Free
- add frontend CPA import flow and account import API helper
- add tests for imported plan type detection and preservation

+++ internal/backend/pow.go
@@ -94,7 +94,7 @@ func buildPOWConfig(userAgent string, scriptSources []string, dataBuild string) 
 	return []any{
 		randomChoiceInt([]int{3000, 4000, 5000}),
 		now,
-		4294705152,
+		int64(4294705152),
 		0,
 		userAgent,
 		randomChoice(scriptSources),
